refactor: Like 이동
- Like와 관련된 클래스들을 like.command.domain.like에서 like.command.domain으로 이동
- import 구문 및 관련 참조 경로 업데이트

#158
