+++ src/main/java/atwoz/atwoz/member/query/introduction/intra/IntroductionQueryRepository.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 import java.util.*;
 
 import static atwoz.atwoz.interview.command.domain.answer.QInterviewAnswer.interviewAnswer;
-import static atwoz.atwoz.like.command.domain.like.QLike.like;
+import static atwoz.atwoz.like.command.domain.QLike.like;
 import static atwoz.atwoz.match.command.domain.match.QMatch.match;
 import static atwoz.atwoz.member.command.domain.introduction.QMemberIntroduction.memberIntroduction;
 import static atwoz.atwoz.member.command.domain.member.QMember.member;
