+++ src/main/java/atwoz/atwoz/member/query/member/MemberQueryRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 
 import static atwoz.atwoz.interview.command.domain.answer.QInterviewAnswer.interviewAnswer;
 import static atwoz.atwoz.interview.command.domain.question.QInterviewQuestion.interviewQuestion;
-import static atwoz.atwoz.like.command.domain.like.QLike.like;
+import static atwoz.atwoz.like.command.domain.QLike.like;
 import static atwoz.atwoz.match.command.domain.match.QMatch.match;
 import static atwoz.atwoz.member.command.domain.member.QMember.member;
 import static atwoz.atwoz.member.command.domain.profileImage.QProfileImage.profileImage;
