+++ src/test/java/atwoz/atwoz/like/command/domain/LikeTest.java
@@ -1,7 +1,5 @@
 package atwoz.atwoz.like.command.domain;
 
-import atwoz.atwoz.like.command.domain.like.Like;
-import atwoz.atwoz.like.command.domain.like.LikeLevel;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
